daemon: Add option --bpf-lb-external-clusterip
Add the option --bpf-lb-external-clusterip to enable
routing to ClusterIP services from outside the cluster.

The loadbalancer routing logic is modified to only look
at the SVC_FLAG_ROUTABLE flag instead of expecting an
additional higher bit in addition to it.

The SVC_FLAG_ROUTABLE is only set for ClusterIP services
when the flag is set.

Fixes: #14581

Signed-off-by: Jussi Maki <jussi@isovalent.com>