egressmap: add support for multiple gateway nodes per policy
This commit lays the groundwork (mostly from a datapath perspective) to
implement high availability egress grateway by adding support for
multiple gateway nodes per policy, sharing all the same egress IP.

The full CFP can be found at [1].

The format of the egress policy map has been updated to allow storing
multiple gateways per policy, and a new CT map has been introduced to
pin connections to a specific gateway.

The CLI logic has also been extended to support managing multiple
gateways for a single policy.

Note that these changes assume the router in front of the gateways will
keep track of which ECMP route a connection belongs to (in order to
forward reply traffic back to the correct gateway).
This in practice is not always true, and in a subsequent PR a more
general approach will be introduced to make sure reply traffic will
eventually hit the correct gateway, regarless of which one the router
forwards the traffic to.

Lastly, for the case of an egress policy with a single gateway, this
change should be backward compatible with the current egress gateway
implementation.

[1] https://docs.google.com/document/d/1ojU8bmy6ai8fdClHCVAPN5-SExm5KjLsh_IJ24tY-4M/edit?usp=sharing

Signed-off-by: Gilberto Bertin <gilberto@isovalent.com>